Added "multisig" feature flag and branched proofs accordingly
An owner / authority can be either an `Account` or a `Multisig` and so
we must conditionally call the appropriate cheatcode based on which case
we would like to prove. Through coniditional compilation we can do this
without having to add other flags into the harnesses, and we only
compile the relevant code for what we are intending to prove.